Completly removed the OWID datasource and made locations static.
Due to invalid data in OWID datasource, which lead to miscalculation, we dropped that source.
There is an official source, made by the EU. We will implement the official source in a later version.
But first, we make the locations table more statically and changed the dataset structure to even hold informations about agegroups and make the calculations more flexible.
This is a major change in code, so all old table structures will no longer work.